requirements.txt
faiss-gpu==1.7.2
pytorch==1.10.0
fairseq==0.12.2
